+++ example/flutter_example/lib/routes/nonlinear_page/nonlinear_data_input.dart
@@ -65,26 +65,27 @@ class _NonlinearDataInputState extends State<NonlinearDataInput> {
     if (formKey.currentState?.validate() ?? false) {
       final precision = context.read<SliderCubit>().state;
       final algorithm = context.read<DropdownCubit>().state;
+      final bloc = context.read<NonlinearBloc>();
 
       if (_getType == NonlinearType.singlePoint) {
-        context.read<NonlinearBloc>().add(
-              SinglePointMethod(
-                method: SinglePointMethod.resolve(algorithm),
-                initialGuess: controllers[1].text,
-                function: controllers[0].text,
-                precision: 1.0 * math.pow(10, -precision),
-              ),
-            );
+        bloc.add(
+          SinglePointMethod(
+            method: SinglePointMethod.resolve(algorithm),
+            initialGuess: controllers[1].text,
+            function: controllers[0].text,
+            precision: 1.0 * math.pow(10, -precision),
+          ),
+        );
       } else {
-        context.read<NonlinearBloc>().add(
-              BracketingMethod(
-                method: BracketingMethod.resolve(algorithm),
-                lowerBound: controllers[1].text,
-                upperBound: controllers[2].text,
-                function: controllers[0].text,
-                precision: 1.0 * math.pow(10, -precision),
-              ),
-            );
+        bloc.add(
+          BracketingMethod(
+            method: BracketingMethod.resolve(algorithm),
+            lowerBound: controllers[1].text,
+            upperBound: controllers[2].text,
+            function: controllers[0].text,
+            precision: 1.0 * math.pow(10, -precision),
+          ),
+        );
       }
     } else {
       ScaffoldMessenger.of(context).showSnackBar(
@@ -117,7 +118,7 @@ class _NonlinearDataInputState extends State<NonlinearDataInput> {
           const SizedBox(height: 40),
 
           // Which algorithm has to be used
-          const DropdownSelection(),
+          const NonlinearDropdownSelection(),
 
           // Some spacing
           const SizedBox(height: 40),
@@ -149,7 +150,7 @@ class _NonlinearDataInputState extends State<NonlinearDataInput> {
                 child: Text(context.l10n.clean),
               ),
             ],
-          )
+          ),
         ],
       ),
     );