+++ example/flutter_example/test/routes/utils/equation_scaffold/bottom_navigation_bar_test.dart
@@ -4,6 +4,7 @@ import 'package:equations_solver/routes/utils/equation_scaffold/navigation_item.
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:golden_toolkit/golden_toolkit.dart';
 
 import '../../mock_wrapper.dart';
 
@@ -27,6 +28,38 @@ void main() {
 
       final bottomNavigation = tester.widget(finder) as BottomNavigation;
       expect(bottomNavigation.navigationItems.length, equals(2));
+    });
+
+    testGoldens('BottomNavigationBar', (tester) async {
+      final builder = GoldenBuilder.column()
+        ..addScenario(
+          'Bottom Navigation Bar',
+          SizedBox(
+            width: 500,
+            height: 150,
+            child: MockWrapper(
+              child: BlocProvider<NavigationCubit>(
+                create: (_) => NavigationCubit(),
+                child: const BottomNavigation(
+                  navigationItems: [
+                    NavigationItem(title: 'Test 1', content: SizedBox()),
+                    NavigationItem(title: 'Test 2', content: SizedBox()),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        );
+
+      await tester.pumpWidgetBuilder(
+        builder.build(),
+        wrapper: (child) => MockWrapper(
+          child: child,
+        ),
+        surfaceSize: const Size(500, 300),
+      );
+
+      await screenMatchesGolden(tester, 'bottom_navigation_bar');
     });
   });
 }