+++ test/system/jacobi_test.dart
@@ -35,6 +35,7 @@ void main() {
       expect(jacobi.maxSteps, equals(30));
       expect(jacobi.precision, equals(1.0e-10));
       expect(jacobi.size, equals(2));
+      expect(jacobi.hasSolution(), isTrue);
 
       // Solutions
       expect(jacobi.determinant(), equals(9));
@@ -42,6 +43,28 @@ void main() {
       final solutions = jacobi.solve();
       expect(solutions[0], const MoreOrLessEquals(7.11, precision: 1.0e-2));
       expect(solutions[1], const MoreOrLessEquals(-3.22, precision: 1.0e-2));
+    });
+
+    test(
+        'Making sure that flat constructor produces the same object that '
+        'a non-flattened constructor does', () {
+      final matrix = JacobiSolver(
+        equations: const [
+          [1, 2, 3],
+          [4, 5, 6],
+          [7, 8, 9],
+        ],
+        constants: const [1, 2, 3],
+        x0: [1, 2, 3],
+      );
+
+      final flatMatrix = JacobiSolver.flatMatrix(
+        equations: const [1, 2, 3, 4, 5, 6, 7, 8, 9],
+        constants: const [1, 2, 3],
+        x0: [1, 2, 3],
+      );
+
+      expect(matrix, equals(flatMatrix));
     });
 
     test('Making sure that the string conversion works properly.', () {
@@ -67,53 +90,74 @@ void main() {
         'Making sure that the matrix is squared AND the dimension of the '
         'known values vector also matches the size of the matrix.', () {
       expect(
-          () => JacobiSolver(equations: const [
-                [1, 2],
-                [4, 5],
-              ], constants: [
-                7,
-                8,
-                9,
-              ], x0: []),
-          throwsA(isA<MatrixException>()));
+        () => JacobiSolver(
+          equations: const [
+            [1, 2],
+            [4, 5],
+          ],
+          constants: [7, 8, 9],
+          x0: [1, 2],
+        ),
+        throwsA(isA<SystemSolverException>()),
+      );
     });
 
     test(
         'Making sure that an exception is thrown when the length of the '
         'initial vector is different from the size of the NxN matrix.', () {
       expect(
-          () => JacobiSolver(equations: const [
-                [1, 2],
-                [4, 5],
-              ], constants: [
-                7,
-                8,
-              ], x0: []),
-          throwsA(isA<SystemSolverException>()));
+        () => JacobiSolver(
+          equations: const [
+            [1, 2],
+            [4, 5],
+          ],
+          constants: [7, 8],
+          x0: [],
+        ),
+        throwsA(isA<SystemSolverException>()),
+      );
+    });
+
+    test(
+        'Making sure that when the input is a flat matrix, the matrix must '
+        'be squared.', () {
+      expect(
+        () => JacobiSolver.flatMatrix(
+          equations: const [1, 2, 3, 4, 5],
+          constants: [7, 8],
+          x0: [1, 2],
+        ),
+        throwsA(isA<MatrixException>()),
+      );
+
+      expect(
+        () => JacobiSolver.flatMatrix(
+          equations: const [1, 2, 3, 4],
+          constants: [7, 8],
+          x0: [1],
+        ),
+        throwsA(isA<SystemSolverException>()),
+      );
     });
 
     test('Making sure that objects comparison works properly.', () {
-      final jacobi = JacobiSolver(equations: const [
-        [1, 2],
-        [3, 4],
-      ], constants: [
-        0,
-        -6,
-      ], x0: [
-        1,
-        2,
-      ]);
+      final jacobi = JacobiSolver(
+        equations: const [
+          [1, 2],
+          [3, 4],
+        ],
+        constants: [0, -6],
+        x0: [1, 2],
+      );
 
-      final jacobi2 = JacobiSolver(equations: const [
-        [1, 2],
-        [3, 4],
-      ], constants: [
-        0,
-        -6,
-      ], x0: [
-        1,
-        2,
-      ]);
+      final jacobi2 = JacobiSolver(
+        equations: const [
+          [1, 2],
+          [3, 4],
+        ],
+        constants: [0, -6],
+        x0: [1, 2],
+      );
 
       expect(jacobi, equals(jacobi2));
       expect(jacobi == jacobi2, isTrue);