+++ test/system/sor_test.dart
@@ -37,6 +37,7 @@ void main() {
       expect(sor.maxSteps, equals(30));
       expect(sor.precision, equals(1.0e-10));
       expect(sor.size, equals(3));
+      expect(sor.hasSolution(), isTrue);
 
       // Solutions
       expect(sor.determinant(), equals(20));
@@ -45,6 +46,28 @@ void main() {
       expect(solutions[0], const MoreOrLessEquals(1, precision: 1.0e-2));
       expect(solutions[1], const MoreOrLessEquals(2, precision: 1.0e-2));
       expect(solutions[2], const MoreOrLessEquals(-2, precision: 1.0e-2));
+    });
+
+    test(
+        'Making sure that flat constructor produces the same object that '
+        'a non-flattened constructor does', () {
+      final matrix = SORSolver(
+        equations: const [
+          [1, 2, 3],
+          [4, 5, 6],
+          [7, 8, 9],
+        ],
+        constants: const [1, 2, 3],
+        w: 1.5,
+      );
+
+      final flatMatrix = SORSolver.flatMatrix(
+        equations: const [1, 2, 3, 4, 5, 6, 7, 8, 9],
+        constants: const [1, 2, 3],
+        w: 1.5,
+      );
+
+      expect(matrix, equals(flatMatrix));
     });
 
     test('Making sure that the string conversion works properly.', () {
@@ -73,33 +96,53 @@ void main() {
         'Making sure that the matrix is squared AND the dimension of the '
         'known values vector also matches the size of the matrix.', () {
       expect(
-          () => SORSolver(equations: const [
-                [1, 2],
-                [4, 5],
-              ], constants: [
-                7,
-                8,
-                9,
-              ], w: 2),
-          throwsA(isA<MatrixException>()));
+        () => SORSolver(
+          equations: const [
+            [1, 2],
+            [4, 5],
+          ],
+          constants: [
+            7,
+            8,
+            9,
+          ],
+          w: 2,
+        ),
+        throwsA(isA<MatrixException>()),
+      );
+    });
+
+    test(
+        'Making sure that when the input is a flat matrix, the matrix must '
+        'be squared.', () {
+      expect(
+        () => SORSolver.flatMatrix(
+          equations: const [1, 2, 3, 4, 5],
+          constants: [7, 8],
+          w: 1.25,
+        ),
+        throwsA(isA<MatrixException>()),
+      );
     });
 
     test('Making sure that objects comparison works properly.', () {
-      final sor = SORSolver(equations: const [
-        [1, 2],
-        [3, 4],
-      ], constants: [
-        0,
-        -6,
-      ], w: 2);
+      final sor = SORSolver(
+        equations: const [
+          [1, 2],
+          [3, 4],
+        ],
+        constants: [0, -6],
+        w: 2,
+      );
 
-      final sor2 = SORSolver(equations: const [
-        [1, 2],
-        [3, 4],
-      ], constants: [
-        0,
-        -6,
-      ], w: 2);
+      final sor2 = SORSolver(
+        equations: const [
+          [1, 2],
+          [3, 4],
+        ],
+        constants: [0, -6],
+        w: 2,
+      );
 
       expect(sor, equals(sor2));
       expect(sor == sor2, isTrue);